+++ libft/ctype.h
@@ -1,0 +1,32 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ctype.h                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mdeville <mdeville@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/01/17 15:06:08 by mdeville          #+#    #+#             */
+/*   Updated: 2018/01/17 15:27:58 by mdeville         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef CTYPE_H
+# define CTYPE_H
+
+int					ft_tolower(int c);
+int					ft_toupper(int c);
+int					ft_isalnum(int c);
+int					ft_isalpha(int c);
+int					ft_isascii(int c);
+int					ft_isblank(int c);
+int					ft_iscntrl(int c);
+int					ft_isdigit(int c);
+int					ft_isgraph(int c);
+int					ft_islower(int c);
+int					ft_isnumber(int c);
+int					ft_isprint(int c);
+int					ft_isspace(int c);
+int					ft_isupper(int c);
+int					ft_isxdigit(int c);
+
+#endif
