+++ libft/ft_io.h
@@ -1,0 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_io.h                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mdeville <mdeville@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/01/17 15:09:23 by mdeville          #+#    #+#             */
+/*   Updated: 2018/01/17 15:28:40 by mdeville         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef FT_IO_H
+# define FT_IO_H
+
+void				ft_putchar(char c);
+void				ft_putchar_fd(char c, int fd);
+void				ft_putendl(char const *s);
+void				ft_putendl_fd(char const *s, int fd);
+void				ft_putnbr(int n);
+void				ft_putnbr_fd(int n, int fd);
+void				ft_putstr(char const *s);
+void				ft_putstr_fd(char const *s, int fd);
+
+int					ft_isatty(int fd);
+
+#endif
