+++ libft/lst.h
@@ -1,0 +1,38 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   lst.h                                              :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mdeville <mdeville@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/01/17 15:11:16 by mdeville          #+#    #+#             */
+/*   Updated: 2018/01/17 15:15:12 by mdeville         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef LST_H
+# define LST_H
+
+#include <unistd.h>
+
+typedef struct		s_list
+{
+	void			*content;
+	size_t			content_size;
+	struct s_list	*next;
+}					t_list;
+
+void		ft_lstadd(t_list **alst, t_list *new);
+void		ft_lstdel(t_list **alst, void (*del)(void *, size_t));
+void		ft_lstdelone(t_list **alst, void (*del)(void *, size_t));
+void		ft_lstfilter(
+						t_list **alst,
+						int (*p)(t_list *),
+						void (*del)(void *, size_t));
+void		ft_lstiter(t_list *lst, void (*f)(t_list *elem));
+t_list		*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
+t_list		*ft_lstnew(void const *content, size_t content_size);
+void		*ft_lstpop(t_list **alst, size_t *content_size);
+void		ft_lstreverse(t_list **alst);
+
+#endif
