+++ libft/memory.h
@@ -1,0 +1,28 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   memory.h                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mdeville <mdeville@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/01/17 15:22:32 by mdeville          #+#    #+#             */
+/*   Updated: 2018/01/17 15:24:16 by mdeville         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef MEMORY_H
+# define MEMORY_H
+
+#include <unistd.h>
+
+void				ft_bzero(void *s, size_t n);
+void				*ft_memalloc(size_t size);
+void				*ft_memccpy(void *dest, const void *src, int c, size_t n);
+void				*ft_memchr(const void *s, int c, size_t n);
+int					ft_memcmp(const void *s1, const void *s2, size_t n);
+void				*ft_memcpy(void *dest, const void *src, size_t n);
+void				ft_memdel(void **ap);
+void				*ft_memmove(void *dest, const void *src, size_t len);
+void				*ft_memset(void *s, int c, size_t n);
+
+#endif
