+++ libft/path.h
@@ -1,0 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   path.h                                             :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mdeville <mdeville@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/01/17 15:18:31 by mdeville          #+#    #+#             */
+/*   Updated: 2018/01/17 15:19:19 by mdeville         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef PATH_H
+# define PATH_H
+
+char	*ft_basename(char *path);
+
+#endif
