+++ libft/string.h
@@ -1,0 +1,63 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   string.h                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mdeville <mdeville@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2018/01/17 15:20:59 by mdeville          #+#    #+#             */
+/*   Updated: 2018/01/17 15:28:58 by mdeville         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef STRING_H
+# define STRING_H
+
+#include <unistd.h>
+# define IS_WS(c) (c == ' ' || c == '\n' || c == '\t')
+
+int					ft_str_is_alpha(const char *str);
+int					ft_str_is_lowercase(const char *str);
+int					ft_str_is_numeric(const char *str);
+int					ft_str_is_printable(const char *str);
+int					ft_str_is_uppercase(const char *str);
+char				*ft_strcapitalize(char *str);
+int					ft_strcasecmp(const char *s1, const char *s2);
+char				*ft_strcat(char *dest, const char *src);
+char				*ft_strchr(const char *s, int c);
+void				ft_strclr(char *s);
+int					ft_strcmp(const char *s1, const char *s2);
+char				*ft_strcpy(char *dest, const char *src);
+void				ft_strdel(char **as);
+char				*ft_strdup(const char *s1);
+int					ft_strequ(char const *s1, char const *s2);
+void				ft_striter(char *s, void (*f)(char *));
+void				ft_striteri(char *s, void (*f)(unsigned int, char *));
+char				*ft_strjoin(char const *s1, char const *s2);
+char				*ft_strnjoin(size_t n, ...);
+size_t				ft_strlcat(char *dst, const char *src, size_t size);
+size_t				ft_strlen(const char *str);
+size_t				ft_strnlen(const char *str, size_t n);
+char				*ft_strmap(char const *s, char (*f)(char));
+char				*ft_strmapi(char const *s, char (*f)(unsigned int, char));
+int					ft_strncasecmp(const char *s1, const char *s2, size_t n);
+char				*ft_strncat(char *dest, const char *src, size_t n);
+int					ft_strncmp(const char *s1, const char *s2, size_t n);
+char				*ft_strncpy(char *dest, const char *src, size_t len);
+char				*ft_strndup(char const *s, size_t n);
+char				*ft_stpncpy(char *dest, const char *src, size_t n);
+int					ft_strnequ(char const *s1, char const *s2, size_t n);
+char				*ft_strnew(size_t size);
+char				*ft_strnstr(
+								const char *big,
+								const char *little,
+								size_t len);
+char				*ft_strrchr(const char *s, int c);
+char				**ft_strsplit(char const *s, char c);
+char				*ft_strstr(const char *haystack, const char *needle);
+char				*ft_strsub(char const *s, unsigned int start, size_t len);
+char				*ft_strtrim(char const *s);
+char				*ft_strtrimc(char const *s, char c);
+int					to_utf8(unsigned int c, char *utf8);
+
+#endif
